requirements.txt
# Night Writer Terminal Automation System Requirements

# Core dependencies
psutil>=5.9.0

# Windows API dependencies (for existing window support)
pywin32>=306; sys_platform == "win32"
Pillow>=10.0.0
easyocr>=1.7.0
pygetwindow>=0.0.9
pyautogui>=0.9.54

# Optional dependencies for enhanced functionality
# schedule>=1.2.0  # For advanced scheduling (if needed)
# colorama>=0.4.6  # For colored terminal output (if needed)

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=1.0.0
